v1-scripts/requirements.txt
numpy==1.17.4
Pillow==6.2.1
scipy==1.1.0
